+++ igseq/summarize.py
@@ -1,9 +1,8 @@
 """
-Make a summary table of antibody attributes, via IgBLAST.
+Make a summary table of antibody attributes via IgBLAST.
 """
 
 import logging
-import subprocess
 from csv import DictReader, DictWriter
 from io import StringIO
 from pathlib import Path
@@ -26,26 +25,25 @@ def summarize(ref_paths, query, output=None, showtxt=None, species=None, dry_run
         showtxt = not output
         LOGGER.info("detected showtxt: %s", showtxt)
     attrs_list = vdj.parse_vdj_paths(ref_paths)
-    species_igblast = igblast.detect_species(attrs_list, species)
+    species_det = {attrs.get("species") for attrs in attrs_list}
+    species_det = {s for s in species_det if s}
+    organism = igblast.detect_organism(species_det, species)
     attrs_list_grouped = vdj.group(attrs_list)
     for key, attrs_group in attrs_list_grouped.items():
         LOGGER.info("detected %s references: %d", key, len(attrs_group))
         if len(attrs_group) == 0:
-            raise util.IgSeqError("No references for segment %s" % key)
+            raise util.IgSeqError(f"No references for segment {key}")
 
     if not dry_run:
         results = []
-        proc = igblast.setup_db_and_igblast(
-            attrs_list_grouped, species_igblast, query, threads=threads,
+        proc, attrs_list_seq = igblast.setup_db_dir_and_igblast(
+            [attrs["path"] for attrs in attrs_list], organism, query, threads=threads,
             extra_args=["-outfmt", "19"],
-            stdout=subprocess.PIPE, text=True)
+            capture_output=True, text=True, check=True)
         reader = DictReader(StringIO(proc.stdout), delimiter="\t")
         lengthmap = {}
-        for attrs_group in attrs_list_grouped.values():
-            for attrs in attrs_group:
-                with open(attrs["path"]) as f_in:
-                    for record in SeqIO.parse(f_in, "fasta"):
-                        lengthmap[record.id] = len(record)
+        for attrs in attrs_list_seq:
+            lengthmap[attrs["seqid_here"]] = len(attrs["seq"])
         for row in reader:
             cdrlens = []
             for cdr in ["cdr1_aa", "cdr2_aa", "cdr3_aa"]: