+++ tools/check-examples.sh
@@ -2,9 +2,11 @@
 
 set -e
 
-export EXAMPLES_HERE=$(readlink -f igseq/data/examples)
+export EXAMPLES=$(readlink -f igseq/data/examples)
 export HERE=$PWD
-for script in "$EXAMPLES_HERE"/*.sh; do
+export PATH_ORIG=$PATH
+export PATH=".:$PATH"
+for script in "$EXAMPLES"/*.sh; do
 	example=$(basename $script)
 	example=${example%.sh}
 	echo $example...
@@ -13,10 +15,8 @@ for script in "$EXAMPLES_HERE"/*.sh; do
 	tmpdir=$(mktemp -d -p examples-tmp ${example}.XXXXXXXX)
 	pushd $tmpdir
 	ln -s $HERE/tools/example-wrapper.sh igseq
-	export PATH_ORIG=$PATH
-	export PATH=".:$PATH"
-	$script
-	rsyncout=$(rsync -crin "$EXAMPLES_HERE/outputs/$example/" .)
+	(set -e; source $script)
+	rsyncout=$(rsync -crin "$EXAMPLES/outputs/$example/" .)
 	if [[ "$rsyncout" != "" ]]; then
 		echo "$rsyncout"
 		exit 1